+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ### Changelog
 ## 1.0.0
 - Initial version
+
+## 1.1.0
+- Added new settings: "Major/Minor/Misc. Skill Cap"
+- Fixed an issue with dependent settings not properly initializing
+- Fixed an issue with settings version detection